Refactor logClick functions into analytics
