+++ amundsen_application/static/js/components/OwnerEditor/index.tsx
@@ -9,9 +9,9 @@ import AppConfig from 'config/config';
 import AvatarLabel, { AvatarLabelProps } from 'components/AvatarLabel';
 import LoadingSpinner from 'components/LoadingSpinner';
 import { ResourceType, UpdateMethod, UpdateOwnerPayload } from 'interfaces';
+import { logClick } from 'utils/analytics';
 
 import { EditableSectionChildProps } from 'components/EditableSection';
-import { logClick } from 'ducks/utilMethods';
 
 import * as Constants from './constants';
