+++ amundsen_application/static/js/components/ScrollTracker/index.tsx
@@ -3,7 +3,7 @@
 
 import * as React from 'react';
 import { throttle } from 'throttle-debounce';
-import { logAction } from 'ducks/utilMethods';
+import { logAction } from 'utils/analytics';
 
 export interface ScrollTrackerProps {
   targetId: string;
