+++ amundsen_application/static/js/ducks/utilMethods.ts
@@ -1,6 +1,4 @@
 import { Tag } from 'interfaces';
-import { ActionLogParams, ClickLogParams, postActionLog } from './log/api/v0';
-import { trackEvent } from '../utils/analytics';
 
 export function sortTagsAlphabetical(a: Tag, b: Tag): number {
   return a.tag_name.localeCompare(b.tag_name);
@@ -28,42 +26,4 @@ export function filterFromObj(
       obj[key] = initialObj[key];
       return obj;
     }, {});
-}
-
-export function logAction(declaredProps: ActionLogParams) {
-  const props = {
-    location: window.location.pathname,
-    ...declaredProps,
-  };
-  postActionLog(declaredProps);
-  trackEvent(declaredProps.command, declaredProps);
-}
-
-export function logClick(
-  event: React.MouseEvent<HTMLElement>,
-  declaredProps?: ClickLogParams
-) {
-  const target = event.currentTarget;
-  const inferredProps: ActionLogParams = {
-    command: 'click',
-    target_id:
-      target.dataset && target.dataset.type ? target.dataset.type : target.id,
-    label: target.innerText || target.textContent || '',
-  };
-
-  if (target.nodeValue !== null) {
-    inferredProps.value = target.nodeValue;
-  }
-
-  let nodeName = target.nodeName.toLowerCase();
-  if (nodeName === 'a') {
-    if (target.classList.contains('btn')) {
-      nodeName = 'button';
-    } else {
-      nodeName = 'link';
-    }
-  }
-  inferredProps.target_type = nodeName;
-
-  logAction({ ...inferredProps, ...declaredProps });
 }
