+++ amundsen_application/static/js/pages/DashboardPage/index.tsx
@@ -10,7 +10,6 @@ import * as ReactMarkdown from 'react-markdown';
 import { getDashboard } from 'ducks/dashboard/reducer';
 import { GetDashboardRequest } from 'ducks/dashboard/types';
 import { GlobalState } from 'ducks/rootReducer';
-import { logAction, logClick } from 'ducks/utilMethods';
 import { UpdateSearchStateRequest } from 'ducks/search/types';
 import { updateSearchState } from 'ducks/search/reducer';
 
@@ -34,6 +33,7 @@ import { getLoggingParams } from 'utils/logUtils';
 
 import { ResourceType } from 'interfaces';
 import { DashboardMetadata } from 'interfaces/Dashboard';
+import { logAction, logClick } from 'utils/analytics';
 import { NO_TIMESTAMP_TEXT } from '../../constants';
 import {
   ADD_DESC_TEXT,
