+++ amundsen_application/static/js/pages/TableDetailPage/constants.ts
@@ -11,7 +11,9 @@ export const TAG_TITLE = 'Tags';
 export const SORT_BY_DROPDOWN_TITLE = 'Sort by';
 export const SORT_BY_MENU_TITLE_TEXT = 'Sort by';
 
-export const COLUMN_TAB_KEY = 'columns';
-export const DASHBOARD_TAB_KEY = 'dashboard';
-export const DOWNSTREAM_TAB_KEY = 'downstream';
-export const UPSTREAM_TAB_KEY = 'upstream';
+export enum TABLE_TAB {
+  COLUMN = 'columns',
+  DASHBOARD = 'dashboard',
+  DOWNSTREAM = 'downstream',
+  UPSTREAM = 'upstream',
+}
